refactor: use `chess.js` internally
In order to support more advanced features, I decided to make use of the
`chess.js` library. This will outsource FEN validation, importing,
chessboard representation, PGN support, history and more. Chess.js is a
more mature library, and it would be a waste of resources to
re-implement most of it. This should make easier to support more
advanced features in the future (such as highlighting all the attacking
pieces, add PGN suppor, add more.